Use a size_t instead of a uint32_t for cancellation flag

+++ lib/binding/lib.rs
@@ -15,7 +15,7 @@ use std::collections::HashMap;
 use std::ffi::CStr;
 use std::marker::PhantomData;
 use std::os::raw::{c_char, c_void};
-use std::sync::atomic::AtomicU32;
+use std::sync::atomic::AtomicUsize;
 use std::{fmt, ptr, slice, str, u16};
 
 pub const PARSER_HEADER: &'static str = include_str!("../include/tree_sitter/parser.h");
@@ -336,13 +336,13 @@ impl Parser {
         };
     }
 
-    pub unsafe fn cancellation_flag(&self) -> Option<&AtomicU32> {
-        (ffi::ts_parser_cancellation_flag(self.0) as *const AtomicU32).as_ref()
+    pub unsafe fn cancellation_flag(&self) -> Option<&AtomicUsize> {
+        (ffi::ts_parser_cancellation_flag(self.0) as *const AtomicUsize).as_ref()
     }
 
-    pub unsafe fn set_cancellation_flag(&self, flag: Option<&AtomicU32>) {
+    pub unsafe fn set_cancellation_flag(&self, flag: Option<&AtomicUsize>) {
         if let Some(flag) = flag {
-            ffi::ts_parser_set_cancellation_flag(self.0, flag as *const AtomicU32 as *const u32);
+            ffi::ts_parser_set_cancellation_flag(self.0, flag as *const AtomicUsize as *const usize);
         } else {
             ffi::ts_parser_set_cancellation_flag(self.0, ptr::null());
         }
